+++ src/components/Login/Login.js
@@ -1,15 +1,16 @@
 import React, { Component } from 'react';
-import ReactDOM from 'react-dom';
+// import ReactDOM from 'react-dom';
 import axios from 'axios';
 import { Link } from 'react-router-dom';
 import './Login.css';
 
-class Create extends Component {
+class Login extends Component {
   constructor() {
     super();
     this.state = {
       username: '',
       password: '',
+      message: '',
     };
   }
   onChange = e => {
@@ -23,17 +24,31 @@ class Create extends Component {
 
     const { username, password } = this.state;
 
-    axios.post('/api/auth/register', { username, password }).then(result => {
-      this.props.history.push('/login');
-    });
+    axios
+      .post('/api/auth/login', { username, password })
+      .then(result => {
+        localStorage.setItem('jwtToken', result.data.token);
+        this.setState({ message: '' });
+        this.props.history.push('/');
+      })
+      .catch(error => {
+        if (error.response.status === 401) {
+          this.setState({ message: 'Login failed. Username or password not match' });
+        }
+      });
   };
 
   render() {
-    const { username, password } = this.state;
+    const { username, password, message } = this.state;
     return (
       <div class="container">
         <form class="form-signin" onSubmit={this.onSubmit}>
-          <h2 class="form-signin-heading">Register</h2>
+          {message !== '' && (
+            <div class="alert alert-warning alert-dismissible" role="alert">
+              {message}
+            </div>
+          )}
+          <h2 class="form-signin-heading">Please sign in</h2>
           <label for="inputEmail" class="sr-only">
             Email address
           </label>
@@ -59,12 +74,18 @@ class Create extends Component {
             required
           />
           <button class="btn btn-lg btn-primary btn-block" type="submit">
-            Register
+            Login
           </button>
+          <p>
+            Not a member?{' '}
+            <Link to="/register">
+              <span class="glyphicon glyphicon-plus-sign" aria-hidden="true" /> Register here
+            </Link>
+          </p>
         </form>
       </div>
     );
   }
 }
 
-export default Create;
+export default Login;
