+++ src/components/Take/Welcome/index.js
