+++ src/index.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
-import '../node_modules/bootstrap/dist/css/bootstrap-theme.min.css';
+// import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
+// import '../node_modules/bootstrap/dist/css/bootstrap-theme.min.css';
 import './index.css';
 import App from './App';
-import registerServiceWorker from './registerServiceWorker';
+import register from './registerServiceWorker';
 import Login from './components/Login';
 import Register from './components/Register';
 
@@ -19,4 +19,4 @@ ReactDOM.render(
   </Router>,
   document.getElementById('root')
 );
-registerServiceWorker();
+register();
